fix bugs and updated comments for betty

+++ support_func2.c
@@ -55,7 +55,7 @@ int print_str(va_list ap)
  */
 int print_int(va_list ap)
 {
-	int x = va_arg(ap, int);
+	long int x = va_arg(ap, int);
 	int z = 0;
 	long int i = 10000000000;
 	int k = 0;
